fix: bugs in repository and ERD

+++ ERASM/src/main/java/com/project/ERASM/repository/EmployeeCertificationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.project.ERASM.entities.EmployeeCertification;
 
 @Repository
-public interface EmployeeCertificationRepository extends JpaRepository<EmployeeCertification, Integer> {
+public interface EmployeeCertificationRepository extends JpaRepository<EmployeeCertification, Long> {
 
 }
